requirements.txt
torch
pandas
numpy
beir==1.0.0
faiss-cpu==1.7.2